+++ app/request/fs.go
@@ -2,4 +2,5 @@ package request
 
 type DirItemsReq struct {
 	Path string `json:"path"`
+	Page uint32 `json:"page"`
 }